+++ client/src/lib/Helpers.js
@@ -13,6 +13,13 @@ var Helpers = {
     }else {
       return []
     }
+  },
+  replacePlaceholders(match,object){
+    if(match.match(/^[a-zA-Z0-9\[\]\.]*$/)){ /* eslint-disable-line */
+      return eval("object."+match)
+    } else{
+      return undefined
+    }
   }
 };