+++ server/src/lib/postgres.js
@@ -9,20 +9,24 @@ Postgres.query=function(connection_name,query){
 
   return Credential.findByName(connection_name)
   .then((config)=>{
-    return new Client(config)
-  })
-  .then((client)=>{
-    return client.connect()
-  })
-  .then((client)=>{
-    client.query(query,function(err,result){
-      logger.debug("["+connection_name+"] Closing connection")
-      client.end()
-      if(err){
-        throw `[${connection_name}] query error ${err.toString()}`
+    return new Promise((resolve,reject)=>{
+      var client
+      try{
+        client = new Client(config)
+        client.connect()
+      }catch(err){
+        reject(`[${connection_name}] connection error ${err.toString()}`)
       }
-      return result.rows
+      client.query(query,function(err,result){
+        logger.debug("["+connection_name+"] Closing connection")
+        client.end()
+        if(err){
+          reject(`[${connection_name}] query error ${err.toString()}`)
+        }
+        resolve(result?.rows)
+      })
     })
+
   })
 };
 module.exports = Postgres